move scripts/config.js to config.js
