+++ config.js
@@ -2,7 +2,11 @@ import path from "path";
 
 const SCAFFOLD_DIRECTORY = path.join("scaffold");
 const DEMO_DIRECTORY = path.join("demo");
+const DJANGO_DEMO_DIRECTORY = path.join("demo", "backend");
+const DEMO_NODE_MODULES_DIRECTORY = path.join("demo", "node_modules");
 const DISTRIBUTION_DIRECTORY = path.join("dist");
+const RAW_DISTRIBUTION_DIRECTORY = path.join("dist", "raw");
+const COOKIE_DISTRIBUTION_DIRECTORY = path.join("dist", "cookie");
 
 export default {
   constants: {},
@@ -13,7 +17,7 @@ export default {
     directory: DISTRIBUTION_DIRECTORY,
     builds: {
       raw: {
-        directory: path.join(DISTRIBUTION_DIRECTORY, "raw"),
+        directory: RAW_DISTRIBUTION_DIRECTORY,
         // Avoids react-native cli error:
         // "Project name shouldn't contain "HelloWorld" name in it,
         // because it is CLI's default placeholder name."
@@ -22,7 +26,7 @@ export default {
         titlePlaceholder: "Hello App Display Name"
       },
       cookie: {
-        directory: path.join(DISTRIBUTION_DIRECTORY, "cookie"),
+        directory: COOKIE_DISTRIBUTION_DIRECTORY,
         config: {
           project_name: "{{cookiecutter.project_name}}",
           project_slug:
@@ -37,6 +41,16 @@ export default {
   },
   demo: {
     directory: DEMO_DIRECTORY,
+    backendDirectory: DJANGO_DEMO_DIRECTORY,
     placeholderName: "demo"
+  },
+  upgrade: {
+    manifest: "upgrade-manifest.js",
+    slugPlaceholder: "{{slug}}",
+    ignoreTemplatize: [
+      "ios/fastlane/metadata/review_information/review_demo_password.txt",
+      "ios/fastlane/metadata/review_information/review_demo_user.txt"
+    ],
+    ignoreDirectories: [DJANGO_DEMO_DIRECTORY, DEMO_NODE_MODULES_DIRECTORY]
   }
 };
